Change calculation of maximum gausian-curve time to be 1.4*MAX
This means that very fast and very slow programs do not crowd the left
or right hand side of the graph. The figures should lie closer (but not
exactly at) the center. Using mean value was considered, but because the
min value is most meaningful, it is hard to tell if centering on mean
has any meaning at all. This way, we at least see the entire tail.